refactor: Image Service -> hwanCache 사용해서 memory, disk 저장
